+++ test/system-test-sentry-server.py
@@ -15,6 +15,27 @@ uri = urlparse(sys.argv[1] if len(sys.argv) > 1 else 'http://127.0.0.1:8000')
 version='1.1.0'
 appIdentifier='com.sentry.fastlane.app'
 
+class EnvelopeStorage:
+    __envelopes_received = []
+
+    @classmethod
+    def add(cls, envelope):
+        cls.__envelopes_received.append(envelope)
+
+    @classmethod
+    def get_envelopes_received(cls):
+        return cls.__envelopes_received
+
+    @classmethod
+    def get_json(cls):
+        jsonObject = {
+            'envelopes': cls.__envelopes_received
+        }
+        return json.dumps(jsonObject)
+
+    @classmethod
+    def reset(cls):
+        cls.__envelopes_received.clear()
 
 class EnvelopeCount:
     __envelopes_received = 0
@@ -34,6 +55,10 @@ class EnvelopeCount:
         }
         return json.dumps(jsonObject)
 
+    @classmethod
+    def reset(cls):
+        cls.__envelopes_received = 0
+
 
 class Handler(BaseHTTPRequestHandler):
     body = None
@@ -49,6 +74,18 @@ class Handler(BaseHTTPRequestHandler):
         if self.path == "/envelope-count":
             print("Envelope count queried " + str(EnvelopeCount.get_envelopes_received()))
             self.writeJSON(EnvelopeCount.get_json())
+            return
+
+        if self.path == "/envelopes-received":
+            print("Envelopes queried ")
+            self.writeJSON(EnvelopeStorage.get_json())
+            return
+
+        if self.path == "/reset":
+            print("Envelopes reset")
+            EnvelopeStorage.reset()
+            EnvelopeCount.reset()
+            self.writeJSON(json.dumps({}))
             return
 
         self.flushLogs()
@@ -76,8 +113,9 @@ class Handler(BaseHTTPRequestHandler):
         if isinstance(code, HTTPStatus):
             code = code.value
         body = self.body = self.requestBody()
- 
+
         if body:
+            EnvelopeStorage.add(str(body))
             body = self.body[0:min(1000, len(body))]
         self.log_message('"%s" %s %s%s',
                          self.requestline, str(code), "({} bytes)".format(len(body)) if size else '', body)
@@ -117,10 +155,10 @@ class Handler(BaseHTTPRequestHandler):
     def getContent(self):
         length = int(self.headers['Content-Length'])
         content = self.rfile.read(length)
-        
+
         if 'Content-Encoding' in self.headers and self.headers['Content-Encoding'] == 'gzip':
             content = gzip.decompress(content)
-        
+
         return content